formatter-service/app/requirements.txt
grpcio
grpcio-tools
Flask
requests